変更

+++ web/app/resources/views/commons/favorite_button.blade.php
@@ -1,5 +1,5 @@
 @if (Auth::id() != $post->id)
-    @if (Auth::user() -> is_favorites($post->id))
+    @if (Auth::user() && Auth::user()->is_favorites($post->id))
         {{--unfavorite--}}
         {!! Form::open(['route' => ['unfavorite', $post->id], 'method' => 'delete']) !!}
             {!! Form::button('<i class="fas fa-star"></i>', ['class' => "btn btn-outline-warning btn-sm", 'type' => 'submit']) !!}
